+++ src/imgui.ini
@@ -1,5 +1,5 @@
 [Window][Debug##Default]
-Pos=60,60
+Pos=7,60
 Size=400,400
 Collapsed=0
 
@@ -34,7 +34,7 @@ Size=928,922
 Collapsed=1
 
 [Window][Main Application]
-Pos=778,78
-Size=737,204
+Pos=7,54
+Size=737,228
 Collapsed=0
